# Final-Project/domain/user.go
package domain

import (
	"MYGRAM-API/helpers"
	"context"
	"time"

	"github.com/asaskevich/govalidator"
	"gorm.io/gorm"
)

type User struct {
	ID              string         `gorm:"primaryKey;type:VARCHAR(50)" json:"id"`
	Username        string         `gorm:"type:VARCHAR(50);uniqueIndex;not null" valid:"required" form:"username" json:"username" example:"sauki"`
	Email           string         `gorm:"type:VARCHAR(50);uniqueIndex;not null" valid:"email,required" form:"email" json:"email" example:"saukiadillah@gmail.com"`
	Password        string         `gorm:"not null" valid:"required,minstringlength(6)" form:"password" json:"password,omitempty" example:"sauki"`
	Age             uint           `gorm:"not null" valid:"required,range(8|63)" form:"age" json:"age,omitempty" example:"8"`
	ProfileImageUrl string         `json:"profileImageUrl,omitempty"`
	CreatedAt       *time.Time     `gorm:"not null;autoCreateTime" json:"created_at,omitempty"`
	UpdatedAt       *time.Time     `gorm:"not null;autocreateTime" json:"updated_at,omitempty"`
	Photos          *[]Photo       `json:"-"`
	SocialMedias    *[]SocialMedia `json:"-"`
}

func (user *User) BeforeCreate(db *gorm.DB) (err error) {
	if _, err := govalidator.ValidateStruct(user); err != nil {
		return err
	}

	user.Password = helpers.Hash(user.Password)

	return
}

func (user *User) BeforeUpdate(db *gorm.DB) (err error) {
	if _, err := govalidator.ValidateStruct(user); err != nil {
		return nil
	}

	return
}

type UserUseCase interface {
	Register(context.Context, *User) error
	Login(context.Context, *User) error
	Update(context.Context, User) (User, error)
	Delete(context.Context, string) error
}

type UserRepository interface {
	Register(context.Context, *User) error
	Login(context.Context, *User) error
	Update(context.Context, User) (User, error)
	Delete(context.Context, string) error
}